Move chunks to extras module.
Refactor accordingly
