+++ iterext/recipes.py
@@ -3,16 +3,14 @@
 """
 
 import collections
-import functools
 from itertools import chain, combinations, count, cycle, groupby, imap
 from itertools import ifilterfalse, islice, izip, izip_longest, repeat, starmap
-from itertools import takewhile, tee
+from itertools import tee
 import operator
 import random
 
 
 __all__ = (
-    'chunks',
     'consume',
     'dotproduct',
     'flatten',
@@ -55,28 +53,6 @@ def flatten(listOfLists):
         [1, 2, 3, 11, 22, 33]
     """
     return chain.from_iterable(listOfLists)
-
-
-def chunks(iterable, size=20):
-    """Generator that iterates over an iterable in size chunks
-
-    >>> list(chunks('ABCDEFG', size=3))
-    [('A', 'B', 'C'), ('D', 'E', 'F'), ('G',)]
-
-    >>> list(chunks([1, 2, 3, 4, 5, 6], size=2))
-    [(1, 2), (3, 4), (5, 6)]
-    """
-    stop = object()
-
-    for chunk in grouper(size, iterable, fillvalue=stop):
-        if chunk[-1] is stop:
-            is_not_stop = functools.partial(operator.is_not, stop)
-
-            yield tuple(takewhile(is_not_stop, chunk))
-
-            break
-
-        yield chunk
 
 
 def grouper(n, iterable, fillvalue=None):
